requirements.txt
keras
tensorflow
Flask==2.2.5
numpy==1.21.5
Pillow==9.5.0
gevent==22.10.2
requests==2.31.0
Werkzeug==2.2.3
h5py==2.10.0
opencv-python==4.8.0.74
